fix: Change retry logic to extract next run from traceback
`task_meta.result` is an Exception for me with Celery v5.2.5, I can't grasp where result.when would come from. The retry logic simply does not work as the result is an Exception and does not have any fields indicating when the next run is.

Still kept it in and checked if it's iterable, maybe it works for others.

Here's a PR that works for me, if there's other solutions let me know. But atm retries are causing 500 errors.